refactor: 拆分前端上帝组件，收敛逻辑到组合式函数与工具模块
- 抽取 usePlayerCore/useDanmaku/useSubtitleTracks/usePlayProgress/useAnimeData，
  Player.vue 缩减约 1300 行；AnimeDetail/MediaLibrary/InstallStep 同步收敛
- 抽取 useMediaLibrary/useServerPagination/usePolling/useMessageActions 等
  后台与安装页共用逻辑
- 登录/注册/消息铃铛抽为 LoginDialog/RegisterDialog/MessageBell 组件
- utils 收敛：siteConfig 缓存读写、bbcode 渲染、playProgress 文案、format 时间格式化
- http 拦截器统一业务错误（BusinessError），会话过期弹窗去重
- 修复：追番状态竞态守卫、SSE 卸载兜底断开、summary/staff HTML 统一 sanitize
- 路由 /admin 增加角色鉴权；远端访问开关解析统一到 siteConfig
- 构建：vite-plugin-vuetify 自动导入 + manualChunks 分包 + 环境变量化代理配置
