+++ ui/.env.development
@@ -1,0 +1,5 @@
+# 开发环境配置
+# VITE_API_TARGET=http://localhost:8081
+VITE_API_TARGET=http://192.168.123.76:8085
+# VITE_IMAGES_TARGET=http://localhost:8081
+VITE_DEV_PORT=5173
