+++ ui/.env.production
@@ -1,0 +1,5 @@
+# 生产环境配置
+# 生产环境默认走同源 /api 与 /images，无需额外配置。
+# 如需部署到独立后端，取消注释并填写对应地址。
+# VITE_API_TARGET=https://api.example.com
+# VITE_IMAGES_TARGET=https://images.example.com
