+++ ui/src/composables/useAnimeDerived.js
@@ -1,10 +1,15 @@
 import { computed } from 'vue'
 import { WEEKDAY_LABELS } from '../utils/constants'
+import { sanitizeHtml } from '../utils/sanitize'
 
 // 从原始 anime 数据（ref）派生的展示字段。
 // 此前这些 computed 在 AnimeDetail.vue / Player.vue 中几乎逐字重复，统一收敛到这里。
 
 const STAFF_META_KEYS = ['原作', '导演', '音乐', '动画制作']
+
+const DEFAULT_RATING_BANGUMI = '7.8'
+const DEFAULT_RATING_ANIDB = '8.51'
+const DEFAULT_TOTAL_EPISODES = '10'
 
 function formatRating(value, digits = 1) {
   const num = Number(value)
@@ -14,17 +19,18 @@ function formatRating(value, digits = 1) {
 export function useAnimeDerived(animeData) {
   const isOnAir = computed(() => Boolean(animeData.value?.isOnAir))
   const ratingMain = computed(() => formatRating(animeData.value?.rating, 1))
-  const ratingBangumi = computed(() => animeData.value?.ratingDetails?.['Bangumi评分'] ?? '7.8')
-  const ratingAnidb = computed(() => animeData.value?.ratingDetails?.['Anidb连载中评分'] ?? '8.51')
+  const ratingBangumi = computed(() => animeData.value?.ratingDetails?.['Bangumi评分'] ?? DEFAULT_RATING_BANGUMI)
+  const ratingAnidb = computed(() => animeData.value?.ratingDetails?.['Anidb连载中评分'] ?? DEFAULT_RATING_ANIDB)
 
   const totalEpisodes = computed(() => {
     const meta = animeData.value?.metadata || []
     const epItem = meta.find((m) => m.startsWith('话数'))
-    return epItem ? epItem.split(':')[1]?.trim() : '10'
+    return epItem ? epItem.split(':')[1]?.trim() : DEFAULT_TOTAL_EPISODES
   })
 
+  // 直接产出 HTML 片段（换行转 <br>），在此处统一清洗，避免调用方忘记 sanitize 造成 XSS 面
   const formattedSummary = computed(() => {
-    return animeData.value?.summary?.replace(/\n/g, '<br>') || ''
+    return sanitizeHtml(animeData.value?.summary?.replace(/\n/g, '<br>') || '')
   })
 
   const titleInfo = computed(() => {
@@ -41,6 +47,7 @@ export function useAnimeDerived(animeData) {
     return WEEKDAY_LABELS[day] || ''
   })
 
+  // staffList 元素本身可能含 <strong> 标签，统一在此清洗
   const staffList = computed(() => {
     const meta = animeData.value?.metadata || []
     return meta
@@ -48,9 +55,9 @@ export function useAnimeDerived(animeData) {
       .map((item) => {
         const parts = item.split(':')
         if (parts.length >= 2) {
-          return `<strong>${parts[0]}:</strong>${parts.slice(1).join(':')}`
+          return sanitizeHtml(`<strong>${parts[0]}:</strong>${parts.slice(1).join(':')}`)
         }
-        return item
+        return sanitizeHtml(item)
       })
   })
