+++ ui/src/main.js
@@ -1,7 +1,5 @@
 import { createApp, watch } from 'vue'
 import { createVuetify } from 'vuetify'
-import * as components from 'vuetify/components'
-import * as directives from 'vuetify/directives'
 import { zhHans } from 'vuetify/locale'
 import 'vuetify/styles'
 import '@mdi/font/css/materialdesignicons.css'
@@ -15,8 +13,6 @@ import { setupHttpInterceptors } from './utils/http'
 import { theme, isDark } from './composables/useTheme'
 
 const vuetify = createVuetify({
-  components,
-  directives,
   locale: {
     locale: 'zh-Hans',
     messages: {
